New Jest Tests and streamlining
The functions that have automated tests were all moved to new file "testedFunctions", and new tests were added with them.
