+++ index.html
@@ -159,5 +159,6 @@
     <script src="JavaScript/calendar.js"></script>
 	  <script src="JavaScript/meals.js"></script>
     <script src="JavaScript/theme.js"></script>
+    <script src="JavaScript/testedFunctions.js"></script>
   </body>
 </html>
