+++ script.test.js
@@ -1,5 +1,7 @@
-const { addTask, removeTask, doesDayHaveTasks } = require(".JavaScript/script"); // Adjust the path accordingly
+const { generateDateString, doesDayHaveTasks, addTask, removeTask, populateUpcomingEvents, fixTime } = require("./JavaScript/testedFunctions.js")
+
 const dateString = "2025-03-22"
+global.setTheme = jest.fn();
 
 beforeEach(() => {
     // Mock the DOM
@@ -9,6 +11,8 @@ beforeEach(() => {
     <input id="location" value="Office" />
     <input id="desc" value="Team Meeting with Frank" />
     <input id="title" value="Meeting" />
+    <div id="upcoming-event-sect"></div>
+    <div id="no-Tasks"></div>
   `;
 
     // Mock localStorage
@@ -24,7 +28,17 @@ beforeEach(() => {
             this.store = {};
         }
     };
+
+    const constantDate = new Date('2025-04-10T00:00:00');
+    jest.useFakeTimers('modern');
+    jest.setSystemTime(constantDate);
+
 });
+
+afterAll(() => {
+    jest.useRealTimers();
+});
+
 
 //ACTUAL TESTS BELOW
 
@@ -87,5 +101,58 @@ test("removeTask function properly deletes a task from localStorage.", () => {
 
     expect(localStorage.getItem("2025-03-22-task1")).toBe(null);
     expect(doesDayHaveTasks(dateString)).toBeTruthy();
+
+});
+
+test('generateDateString produces correct strings', () => {
+    const dateString = generateDateString(10, true);
+    expect(dateString).toMatch("2025-04-10");
+
+});
+
+
+test("populateUpcomingEvents edits the proper div and hides no-Tasks div", () => {
+    localStorage.clear();
+
+
+    localStorage.setItem("2025-04-10-taskAmount", "1");
+    localStorage.setItem("2025-04-10-taskStart", "1");
+    localStorage.setItem("2025-04-10-task1", "Math Review");
+    localStorage.setItem("2025-04-10-time1", "1:00 pm");
+    localStorage.setItem("2025-04-10-addy1", "Library");
+    localStorage.setItem("2025-04-10-desc1", "Study session");
+
+    // Run function
+    populateUpcomingEvents();
+
+    // Check if task was appended
+    const tasks = document.querySelectorAll(".upcoming-event");
+    expect(tasks.length).toBe(1);
+    expect(tasks[0].textContent).toContain("Review");
+    expect(tasks[0].textContent).toContain("1:00 pm");
+
+    // Check if "no-Tasks" div was cleared
+    expect(document.getElementById("no-Tasks").textContent).toBe("");
+
+    removeTask("2025-04-10", 1);
+
+    populateUpcomingEvents();
+
+    expect(document.getElementById("no-Tasks").textContent).toBe("Nothing to do...");
+
+
+});
+
+test("fixTime correctly turns the 24-hour format into 12-hour format.", () => {
+    let time = fixTime(document.getElementById("time").value);
+    expect(time).toBe("5:30 pm");
+
+    document.getElementById("time").value = "1:30"
+    time = fixTime(document.getElementById("time").value);
+    expect(time).toBe("1:30 am");
+
+    document.getElementById("time").value = "13:30"
+    time = fixTime(document.getElementById("time").value);
+    expect(time).toBe("1:30 pm");
 
 });
